Deduce workflow when creating a story
